Move InMemory Reader and Writer to main column-store
